Fixes networking issues, some minor typos
Both commands in the README should work. The main issue was not a docker networking issue (surprisingly), but rather a conf in the test/mainnet containers that block non-local networking access (non-surprising, from a security perspective).

As a dev user, you should be able to run the README commands and connect to the interface server with OCI runtime exec failed: exec failed: container_linux.go:380: starting container process caused: exec: "base": executable file not found in $PATH: unknown and successfully reach the algo node  with  and get a response